requirements.txt
Flask==2.2.2
Flask_Login==0.6.2
flask_sqlalchemy==3.0.3
python-dotenv==1.0.0
requests==2.28.2
gunicorn==20.1.0